created accounts, and account to shard mappings
created accounts and a mapping of accounts to shards which is just random for now. Then for each transaction I randomly add a sender and receiver account. Then determine if it is a cross shard or intra shard transaction. If it is then proceed normally and if not then call the handle cross shard transaction function. Which still needs to be implemented